feat: moves Alert section
