+++ packages/core/app/components/common/Alert/Alert.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import type { PropsWithChildren, ReactNode } from 'react'
 import { useCallback, useState } from 'react'
