+++ packages/core/app/components/common/Alert/section.module.scss
