+++ packages/core/src/components/cms/GlobalSections.tsx
@@ -8,7 +8,7 @@ import { PageContentType, getPage } from 'src/server/cms'
 import Toast from 'src/components/common/Toast'
 import RenderSections from '../../../app/components/cms/RenderSections'
 
-import { OverriddenDefaultAlert as Alert } from 'src/components/sections/Alert/OverriddenDefaultAlert'
+import { OverriddenDefaultAlert as Alert } from 'app/components/sections/Alert/OverriddenDefaultAlert'
 import Footer from 'src/components/sections/Footer'
 import { OverriddenDefaultNavbar as Navbar } from 'src/components/sections/Navbar/OverriddenDefaultNavbar'
 import { OverriddenDefaultRegionBar as RegionBar } from 'src/components/sections/RegionBar/OverriddenDefaultRegionBar'
